+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `submissionTitle` to `Form` type
 
+### Removed
+
+- legacy receipt generation and personalisation code
+
 ## [3.4.0] - 2023-07-26
 
 ### Added